Corrected ProductDeleteCommandHandler; Deleted unused List component; updated api_interfaces.ts with LoginDetails interface

+++ ShoesAndBlouse.Application/Products/CommandHandlers/DeleteProductCommandHandler.cs
@@ -23,7 +23,7 @@ namespace ShoesAndBlouse.Application.Products.CommandHandlers
 
             // Usuń powiązane zdjęcie produktu
             var photoPath = productToDelete.PhotoPath;
-            if (!string.IsNullOrEmpty(photoPath))
+            if (!string.IsNullOrEmpty(photoPath) && File.Exists(photoPath))
             {
                 // Usuń plik zdjęcia
                 File.Delete(photoPath);
